# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cafe Supply Chain API",
        "description": "REST API documentation split by platform: Admin Dashboard, Cafe Mobile App, and Public Storefront.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/cafe/orders": {
            "get": {
                "tags": [
                    "Cafe Mobile Orders"
                ],
                "summary": "List cafe orders",
                "operationId": "df333eb5fb03206b34e826eea2c6b5be",
                "responses": {
                    "200": {
                        "description": "Paginated list of cafe orders"
                    }
                }
            },
            "post": {
                "tags": [
                    "Cafe Mobile Orders"
                ],
                "summary": "Create an order for the cafe",
                "operationId": "4c7d655b789460acf5478b0a6871bed7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CafeOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/cafe/branches/{id}/orders": {
            "get": {
                "tags": [
                    "Cafe Mobile Branches"
                ],
                "summary": "List orders for a specific branch",
                "operationId": "59df518569219287be1f17d7ff72c78d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of branch orders"
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                }
            }
        },
        "/cafe/orders/{id}": {
            "get": {
                "tags": [
                    "Cafe Mobile Orders"
                ],
                "summary": "Get a cafe order",
                "operationId": "ddb54b10046cd006c29ffb454b880dfe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/cafe/orders/{id}/status": {
            "put": {
                "tags": [
                    "Cafe Mobile Orders"
                ],
                "summary": "Update order status",
                "operationId": "766c6aac15f0b1777050785478dff042",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/cafe/branches": {
            "get": {
                "tags": [
                    "Cafe Mobile Branches"
                ],
                "summary": "List cafe branches",
                "operationId": "29bf3a3c3bb631e6885ee6f65afb6f0d",
                "responses": {
                    "200": {
                        "description": "Paginated list of cafe branches"
                    }
                }
            },
            "post": {
                "tags": [
                    "Cafe Mobile Branches"
                ],
                "summary": "Create a branch for the cafe",
                "operationId": "1a40ff98a18f993c5a30a7c0113c49b4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CafeBranchRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Branch created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/cafe/branches/{id}": {
            "get": {
                "tags": [
                    "Cafe Mobile Branches"
                ],
                "summary": "Get a cafe branch",
                "operationId": "8c9baf70781a9927d0ab8dd263b36ba2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branch details"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cafe Mobile Branches"
                ],
                "summary": "Update a cafe branch",
                "operationId": "faa28d4c68cdbaa8d23976bcc1383504",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CafeBranchRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Branch updated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/cafe/profile": {
            "get": {
                "tags": [
                    "Cafe Mobile Profile"
                ],
                "summary": "Get cafe profile",
                "operationId": "b61a587e42e4a65a611e9ce586ebe1f0",
                "responses": {
                    "200": {
                        "description": "Cafe profile"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cafe Mobile Profile"
                ],
                "summary": "Update cafe profile",
                "operationId": "6d026f2b75cdcbd36744e80489abf60f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "contact_info": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated"
                    }
                }
            }
        },
...
        "/cafe/products": {
            "get": {
                "tags": [
                    "Cafe Mobile Catalog"
                ],
                "summary": "List products",
                "operationId": "46d95a00fca533270cd7255284e834f8",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of products"
                    }
                }
            }
        },
        "/cafe/products/{id}": {
            "get": {
                "tags": [
                    "Cafe Mobile Catalog"
                ],
                "summary": "Get product details",
                "operationId": "7cdda903aaa51527f8fd6dbf56255e3b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/cafe/products/{id}/variants": {
            "get": {
                "tags": [
                    "Cafe Mobile Catalog"
                ],
                "summary": "Get product variants by product id",
                "operationId": "0386483cda431404339ea7082d639ecc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of product variants"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CafeBranchRequest": {
                "required": [
                    "cafe_id",
                    "name",
                    "latitude",
                    "longitude"
                ],
                "properties": {
                    "cafe_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "city": {
                        "type": "string",
                        "maxLength": 100,
                        "nullable": true
                    },
                    "street": {
                        "type": "string",
                        "maxLength": 200,
                        "nullable": true
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float"
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float"
                    },
                    "delivery_zone_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "OrderItemRequest": {
                "required": [
                    "product_variant_id",
                    "quantity",
                    "unit_price"
                ],
                "properties": {
                    "product_variant_id": {
                        "type": "integer"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "unit_price": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
...
            "CafeOrderRequest": {
                "required": [
                    "branch_id",
                    "items"
                ],
                "properties": {
                    "branch_id": {
                        "type": "integer"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItemRequest"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}